chore(ci): harden gates this batch actually hit
- Derive remnic-server files[] from src/*.ts.
- Share a regex-free frontmatter splitter for the review deck.
- Make the PR-create stagger test survive a second-boundary tick.
