+++ .changeset/ci-this-run-5.md
@@ -1,0 +1,5 @@
+---
+"@remnic/core": patch
+---
+
+chore(ci): harden gates this batch actually hit
